requirements.txt
requests
pyyaml
jsonschema
pypdf
pymupdf
beautifulsoup4
pytesseract
langdetect
pdf2image
sentence-transformers
numpy
tqdm
redis
python-dotenv
chromadb
faiss-cpu
ollama
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-multipart
jose
passlib[bcrypt]
httpx
authlib
itsdangerous
